feat: Allow override of isSupportingMultipleReports property
In environments where Spotbugs is loaded via classpath instead of a
repository access the mechanism for discovering the version fails to
produce the desired value for the isSupportingMultipleReports property.
Instead of relying on the automatically observed version number this
change makes an optional settable flag that can serve the same purpose
trusting the consumer to make the right choice.