build: GHA Only run verify class files on ubuntu
github didn't add path on windows, this item is not super important IMO as gradle should compile correctly  so there is no real reason to check it worked properly.  However, will keep it for now but limiting it to just one OS as checking numerous times is again IMO worthless.