Additionally run on macos and only run javap on linux as not really needed (#1120)
* build: GHA Echo the path to see if java is property setup

* build: GHA Add macos-latest to build os

* build: GHA Quote echo's as used on windows too

* build: GHA Only run verify class files on ubuntu

github didn't add path on windows, this item is not super important IMO as gradle should compile correctly  so there is no real reason to check it worked properly.  However, will keep it for now but limiting it to just one OS as checking numerous times is again IMO worthless.