feat: Add an extension property to decouple tasks from check
The new runOnCheck extension property (defaults to true) can be
set to false, which disables the dependency from the `check` task
to the spotbugs tasks. This is useful in cases where consumers
do not want it to run automatically as part of `gradle check` but
will run it in a separate context.

Fixes #1303